requirements.txt
# KidsKlassiks FastAPI + HTMX Application Dependencies

# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0

# Session and security middleware
itsdangerous==2.1.2
cryptography==41.0.8

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# AI Services
openai==1.3.7
google-cloud-aiplatform==1.38.1

# HTTP and async requests
aiohttp==3.9.1
requests==2.31.0
httpx==0.25.2

# Additional async dependencies
anyio==4.1.0
sniffio==1.3.0

# Text processing and parsing
beautifulsoup4==4.12.2
markdown==3.5.1
lxml==4.9.3

# PDF generation and document processing
reportlab==4.0.7
weasyprint==60.2
fpdf2==2.7.6

# Image processing
pillow==10.1.0
pdf2image==1.16.3

# Data processing and utilities
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2

# Additional FastAPI dependencies
email-validator==2.1.0
orjson==3.9.10
ujson==5.8.0

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Additional utilities
pathlib2==2.3.7
typing-extensions==4.8.0
pydantic==2.5.0
pydantic-core==2.14.1

# Additional required dependencies
click==8.1.7
h11==0.14.0
idna==3.6
certifi==2023.11.17
charset-normalizer==3.3.2
urllib3==2.1.0
